+++ client/components/templates/TabNew.tsx
@@ -1,17 +1,28 @@
-import React from 'react';
-import { List, Icon } from 'semantic-ui-react';
+import React, { useEffect, useState } from 'react';
+import { List, Icon, Dimmer, Loader } from 'semantic-ui-react';
 import styled from 'styled-components';
 import { useHistory } from 'react-router-dom';
 
 import { Post } from '../../types/post';
+import PostService from '../../repository/post';
 
 type Props = {
-  data: Array<Post>;
+  count: number;
 };
 
 const TabNew = (props: Props) => {
-  const { data } = props;
+  const { count } = props;
   const history = useHistory();
+
+  const [data, setData] = useState();
+
+  useEffect(() => {
+    const postFunc = async () => {
+      const res = await PostService.getPosts(count, 'new');
+      setData(res.data.posts);
+    };
+    postFunc();
+  }, [count]);
 
   const onClickToDetail = () => {
     history.push('/detail');
@@ -31,15 +42,21 @@ const TabNew = (props: Props) => {
                   </List.Description>
                   <List.Description style={{ marginTop: '2px' }}>
                     <Icon name="thumbs down" color="grey" />
-                    ヒドイイネ {obj.favorite_count}　<Icon name="comment" color="grey" />
-                    コメント {obj.comment_count}
+                    ヒドイイネ {obj.favorites_count}　<Icon name="comment" color="grey" />
+                    コメント {obj.comments_count}
                   </List.Description>
                 </List.Content>
               </List.Item>
             );
           })}
         </List>
-      ) : null}
+      ) : (
+        <SCdimmerContainer>
+          <Dimmer active inverted>
+            <Loader inverted></Loader>
+          </Dimmer>
+        </SCdimmerContainer>
+      )}
     </SCcontainer>
   );
 };
@@ -48,6 +65,10 @@ const SCcontainer = styled.div`
   &:hover {
     cursor: pointer;
   }
+`;
+
+const SCdimmerContainer = styled.div`
+  height: 200px;
 `;
 
 export default TabNew;
